feat: 统一Pact管理页面 + CAW chain_id修正(BASE_ETH/TBASE_SETH)
- 新增 /api/pact/status 接口，一次返回所有scope的Pact状态
- 重写 /dashboard/pact 页面为统一Pact管理（Venice + BlockRun）
- Wallet页面和BlockRun页面移除各自的Pact子组件
- CAW chain_id从eip155格式改为CAW原生格式(BASE_ETH/TBASE_SETH)
- 修复Sidebar合并冲突标记、Prisma schema加blockrun_x402枚举
- 修复TypeScript类型错误(blockrun/topup, r34-sweep-heartbeat)

+++ lib/r34-sweep-heartbeat.ts
@@ -197,7 +197,7 @@ async function readCawWalletUsdcBalance(): Promise<number | null> {
     });
 
     // USDC has 6 decimals
-    return Number(formatUnits(balance, 6));
+    return Number(formatUnits(balance as bigint, 6));
   } catch (err) {
     console.warn("[blockrun-balance] viem balance check failed:", err instanceof Error ? err.message : err);
     return null;
